gnhf 41: Extracted the four byte-identical Web Crypto primitives (b64urlEncode/Decode, hmac, timingSafeEqual) into a shared lib/webcrypto.ts, collapsing the duplication between signed-token.ts and admin/gate-token.ts.
