+++ src/lib/webcrypto.ts
@@ -1,0 +1,42 @@
+/**
+ * Shared low-level Web Crypto primitives (Security). PURE + isomorphic: works in
+ * Node and the Edge/middleware runtime. Backs the HMAC-signed token modules
+ * (signed-token.ts for pull/quiz offers, admin/gate-token.ts for the admin gate).
+ */
+
+/** Encode bytes as unpadded base64url. */
+export function b64urlEncode(bytes: Uint8Array): string {
+  let bin = "";
+  for (const b of bytes) bin += String.fromCharCode(b);
+  return btoa(bin).replace(/\+/g, "-").replace(/\//g, "_").replace(/=+$/, "");
+}
+
+/** Decode an unpadded base64url string back to bytes. */
+export function b64urlDecode(s: string): Uint8Array {
+  const pad = s.length % 4 === 0 ? "" : "=".repeat(4 - (s.length % 4));
+  const bin = atob(s.replace(/-/g, "+").replace(/_/g, "/") + pad);
+  const out = new Uint8Array(bin.length);
+  for (let i = 0; i < bin.length; i++) out[i] = bin.charCodeAt(i);
+  return out;
+}
+
+/** HMAC-SHA256 of `payload` under `secret`, as raw bytes. */
+export async function hmac(payload: string, secret: string): Promise<Uint8Array> {
+  const key = await crypto.subtle.importKey(
+    "raw",
+    new TextEncoder().encode(secret),
+    { name: "HMAC", hash: "SHA-256" },
+    false,
+    ["sign"],
+  );
+  const sig = await crypto.subtle.sign("HMAC", key, new TextEncoder().encode(payload));
+  return new Uint8Array(sig);
+}
+
+/** Constant-time byte comparison (no early return on mismatch). */
+export function timingSafeEqual(a: Uint8Array, b: Uint8Array): boolean {
+  if (a.length !== b.length) return false;
+  let diff = 0;
+  for (let i = 0; i < a.length; i++) diff |= a[i] ^ b[i];
+  return diff === 0;
+}
